Decouple DeleteDialog in MeasurementCategoryForm from Modal

+++ shared/src/commonMain/kotlin/com/faltenreich/diaguard/measurement/category/form/MeasurementCategoryFormViewModel.kt
@@ -1,20 +1,12 @@
 package com.faltenreich.diaguard.measurement.category.form
 
 import com.faltenreich.diaguard.measurement.category.MeasurementCategory
-import com.faltenreich.diaguard.navigation.modal.CloseModalUseCase
-import com.faltenreich.diaguard.navigation.modal.OpenModalUseCase
 import com.faltenreich.diaguard.navigation.screen.PopScreenUseCase
 import com.faltenreich.diaguard.shared.architecture.ViewModel
 import com.faltenreich.diaguard.shared.di.inject
-import com.faltenreich.diaguard.shared.localization.Localization
-import com.faltenreich.diaguard.shared.view.AlertModal
-import com.faltenreich.diaguard.shared.view.DeleteModal
-import diaguard.shared.generated.resources.Res
-import diaguard.shared.generated.resources.delete_error_property
-import diaguard.shared.generated.resources.delete_title
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.launch
+import kotlinx.coroutines.flow.combine
+import kotlinx.coroutines.flow.update
 
 class MeasurementCategoryFormViewModel(
     categoryId: Long,
@@ -23,9 +15,6 @@ class MeasurementCategoryFormViewModel(
     private val updateCategory: UpdateMeasurementCategoryUseCase = inject(),
     private val deleteCategory: DeleteMeasurementCategoryUseCase = inject(),
     private val popScreen: PopScreenUseCase = inject(),
-    private val openModal: OpenModalUseCase = inject(),
-    private val closeModal: CloseModalUseCase = inject(),
-    private val localization: Localization = inject(),
 ) : ViewModel<MeasurementCategoryFormState, MeasurementCategoryFormIntent, Unit>() {
 
     val category: MeasurementCategory.Local = checkNotNull(getCategoryBdId(categoryId))
@@ -35,13 +24,24 @@ class MeasurementCategoryFormViewModel(
     var isActive = MutableStateFlow(category.isActive)
 
     private val properties = getProperties(category)
+    private val deleteDialog = MutableStateFlow<MeasurementCategoryFormState.DeleteDialog?>(null)
+    private val alertDialog = MutableStateFlow<MeasurementCategoryFormState.AlertDialog?>(null)
 
-    override val state = properties.map(::MeasurementCategoryFormState)
+    override val state = combine(
+        properties,
+        deleteDialog,
+        alertDialog,
+        ::MeasurementCategoryFormState,
+    )
 
-    override suspend fun handleIntent(intent: MeasurementCategoryFormIntent) {
-        when (intent) {
-            is MeasurementCategoryFormIntent.UpdateCategory -> updateCategory()
-            is MeasurementCategoryFormIntent.DeleteCategory -> deleteCategory()
+    override suspend fun handleIntent(intent: MeasurementCategoryFormIntent) = with(intent) {
+        when (this) {
+            is MeasurementCategoryFormIntent.Store -> updateCategory()
+            is MeasurementCategoryFormIntent.Delete -> deleteCategory(needsConfirmation)
+            is MeasurementCategoryFormIntent.OpenDeleteDialog -> deleteDialog.update { MeasurementCategoryFormState.DeleteDialog }
+            is MeasurementCategoryFormIntent.CloseDeleteDialog -> deleteDialog.update { null }
+            is MeasurementCategoryFormIntent.OpenAlertDialog -> alertDialog.update { MeasurementCategoryFormState.AlertDialog }
+            is MeasurementCategoryFormIntent.CloseAlertDialog -> alertDialog.update { null }
         }
     }
 
@@ -56,28 +56,16 @@ class MeasurementCategoryFormViewModel(
         popScreen()
     }
 
-    private fun deleteCategory() = scope.launch {
+    private suspend fun deleteCategory(needsConfirmation: Boolean) {
         if (category.isUserGenerated) {
-            openModal(
-                DeleteModal(
-                    onDismissRequest = { scope.launch { closeModal() } },
-                    onConfirmRequest = {
-                        deleteCategory(category)
-                        scope.launch {
-                            closeModal()
-                            popScreen()
-                        }
-                    }
-                )
-            )
+            if (needsConfirmation) {
+                deleteDialog.update { MeasurementCategoryFormState.DeleteDialog }
+            } else {
+                deleteCategory(category)
+                popScreen()
+            }
         } else {
-            openModal(
-                AlertModal(
-                    onDismissRequest = { scope.launch { closeModal() } },
-                    title = localization.getString(Res.string.delete_title),
-                    text = localization.getString(Res.string.delete_error_property),
-                )
-            )
+            alertDialog.update { MeasurementCategoryFormState.AlertDialog }
         }
     }
 }
